perf: 优化系统设置模型

+++ sys_controller/sys_settings.go
@@ -6,8 +6,13 @@ import (
 	"github.com/SupenBysz/gf-admin-community/api_v1/sys_api"
 	"github.com/SupenBysz/gf-admin-community/sys_model"
 	"github.com/SupenBysz/gf-admin-community/sys_model/sys_dao"
+	"github.com/SupenBysz/gf-admin-community/sys_model/sys_do"
+	"github.com/SupenBysz/gf-admin-community/sys_model/sys_entity"
 	"github.com/SupenBysz/gf-admin-community/sys_service"
+	"github.com/gogf/gf/v2/frame/g"
+	"github.com/gogf/gf/v2/util/gconv"
 	"github.com/kysion/base-library/base_model"
+	"github.com/kysion/base-library/utility/daoctl"
 )
 
 var SysSettings = cSysSettings{}
@@ -17,11 +22,41 @@ type cSysSettings struct{}
 func (c *cSysSettings) QueryList(ctx context.Context, req *sys_api.QuerySettingListReq) (*sys_model.SysSettingListRes, error) {
 	user := sys_service.SysSession().Get(ctx).JwtClaimsUser
 
-	req.Filter = append(req.Filter, base_model.FilterInfo{
-		Field: sys_dao.SysSettings.Columns().UnionMainId,
-		Where: "=",
-		Value: user.UnionMainId,
-	})
+	queryUnionMainId := user.UnionMainId
+
+	filter := make([]base_model.FilterInfo, 0)
+
+	for i, info := range req.Filter {
+		if info.Field == sys_dao.SysSettings.Columns().UnionMainId {
+			queryUnionMainId = info.Value.(int64)
+		} else {
+			filter = append(filter, req.Filter[i])
+		}
+	}
+
+	cfgValue, err := g.Cfg().Get(ctx, "service.superAdminMainId", 0)
+
+	if cfgValue != nil && err == nil {
+		superAdminMainId := cfgValue.Int64()
+
+		if user.UnionMainId != superAdminMainId {
+			queryUnionMainId = user.UnionMainId
+		} else {
+			if queryUnionMainId == -1 {
+				queryUnionMainId = 0
+			}
+		}
+	}
+
+	if queryUnionMainId > 0 {
+		filter = append(filter, base_model.FilterInfo{
+			Field: sys_dao.SysSettings.Columns().UnionMainId,
+			Where: "=",
+			Value: queryUnionMainId,
+		})
+	}
+
+	req.Filter = filter
 
 	ret, err := sys_service.SysSettings().QueryList(ctx, &req.SearchParams, false)
 
@@ -31,11 +66,38 @@ func (c *cSysSettings) QueryList(ctx context.Context, req *sys_api.QuerySettingL
 func (c *cSysSettings) GetByName(ctx context.Context, req *sys_api.GetSettingByNameReq) (*sys_model.SysSettingsRes, error) {
 	user := sys_service.SysSession().Get(ctx).JwtClaimsUser
 
-	req.Filter = append(req.Filter, base_model.FilterInfo{
-		Field: sys_dao.SysSettings.Columns().UnionMainId,
-		Where: "=",
-		Value: user.UnionMainId,
-	})
+	queryUnionMainId := user.UnionMainId
+	filter := make([]base_model.FilterInfo, 0)
+	for i, info := range req.Filter {
+		if info.Field == sys_dao.SysSettings.Columns().UnionMainId {
+			queryUnionMainId = gconv.Int64(info.Value)
+		} else {
+			filter = append(filter, req.Filter[i])
+		}
+	}
+
+	cfgValue, err := g.Cfg().Get(ctx, "service.superAdminMainId", 0)
+	if cfgValue != nil && err == nil {
+		superAdminMainId := cfgValue.Int64()
+
+		if user.UnionMainId != superAdminMainId && queryUnionMainId != -1 {
+			queryUnionMainId = user.UnionMainId
+		} else {
+			if queryUnionMainId == -1 {
+				queryUnionMainId = 0
+			}
+		}
+	}
+
+	if queryUnionMainId > 0 {
+		filter = append(filter, base_model.FilterInfo{
+			Field: sys_dao.SysSettings.Columns().UnionMainId,
+			Where: "=",
+			Value: queryUnionMainId,
+		})
+	}
+
+	req.Filter = filter
 
 	ret, err := sys_service.SysSettings().GetByName(ctx, req.Name, &req.SearchParams)
 
@@ -44,7 +106,20 @@ func (c *cSysSettings) GetByName(ctx context.Context, req *sys_api.GetSettingByN
 
 func (c *cSysSettings) Save(ctx context.Context, req *sys_api.SaveSettingReq) (*sys_model.SysSettingsRes, error) {
 	user := sys_service.SysSession().Get(ctx).JwtClaimsUser
-	req.UnionMainId = user.UnionMainId
+
+	cfgValue, err := g.Cfg().Get(ctx, "service.superAdminMainId", 0)
+
+	if cfgValue != nil && err == nil {
+		superAdminMainId := cfgValue.Int64()
+
+		if user.UnionMainId != superAdminMainId || req.UnionMainId == 0 {
+			req.UnionMainId = user.UnionMainId
+		} else {
+			if req.UnionMainId == -1 {
+				req.UnionMainId = 0
+			}
+		}
+	}
 
 	ret, err := sys_service.SysSettings().Create(ctx, &req.SysSettings)
 
@@ -53,6 +128,12 @@ func (c *cSysSettings) Save(ctx context.Context, req *sys_api.SaveSettingReq) (*
 
 func (c *cSysSettings) Delete(ctx context.Context, req *sys_api.DeleteSettingReq) (api_v1.BoolRes, error) {
 	user := sys_service.SysSession().Get(ctx).JwtClaimsUser
+
+	selectInfo, err := daoctl.ScanWithError[sys_entity.SysFrontSettings](sys_dao.SysFrontSettings.Ctx(ctx).Where(sys_do.SysFrontSettings{Name: req.Name, UnionMainId: user.UnionMainId}))
+	if selectInfo != nil && selectInfo.UnionMainId <= 0 && user.IsSuperAdmin != true {
+		return false, sys_service.SysLogs().ErrorSimple(ctx, err, "平台配置禁止删除", sys_dao.SysFrontSettings.Table())
+	}
+
 	ret, err := sys_service.SysSettings().Delete(ctx, req.Name, user.UnionMainId)
 
 	return ret == true, err
